completed the all the api except review aand rating  with basic auth

+++ src/main/java/com/mstore/model/Rating.java
@@ -9,9 +9,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Entity
 @Data
@@ -32,7 +30,7 @@ public class Rating {
 	private Product product;
 	
 	@Column(name = "RATING")
-	private float rating;
+	private Double rating;
 	
 
 }
